production_app.py
"""
Production-ready Flask Application for Research Knowledge Graph
Optimized for deployment on Render
"""

import os
import json
import logging
from flask import Flask, render_template, jsonify, request, send_from_directory
from flask_cors import CORS
from flask_compress import Compress
from werkzeug.utils import secure_filename
from werkzeug.middleware.proxy_fix import ProxyFix
import pandas as pd
import networkx as nx
from datetime import datetime
from typing import Dict, List, Any, Optional
import re
from collections import defaultdict
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Configure logging for production
logging.basicConfig(
    level=os.getenv('LOG_LEVEL', 'INFO'),
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger(__name__)

# Initialize Flask app with production settings
app = Flask(__name__)

# Production configuration
app.config.update(
    SECRET_KEY=os.getenv('SECRET_KEY', os.urandom(24)),
    MAX_CONTENT_LENGTH=int(os.getenv('MAX_CONTENT_LENGTH', 16 * 1024 * 1024)),
    UPLOAD_FOLDER=os.getenv('UPLOAD_FOLDER', 'uploads'),
    ALLOWED_EXTENSIONS=set(os.getenv('ALLOWED_EXTENSIONS', 'xlsx,xls,csv').split(',')),
    JSONIFY_MIMETYPE='application/json',
    JSON_SORT_KEYS=False,
    SEND_FILE_MAX_AGE_DEFAULT=31536000,  # 1 year for static files
)

# Security headers and proxy fix for Render
app.wsgi_app = ProxyFix(app.wsgi_app, x_for=1, x_proto=1, x_host=1, x_prefix=1)

# Enable CORS with production settings
CORS(app, origins=os.getenv('CORS_ORIGINS', '*').split(','))

# Enable compression
Compress(app)

# Ensure required directories exist
for directory in ['uploads', 'static/data', 'backups', 'templates']:
    os.makedirs(directory, exist_ok=True)

# Import the EnhancedResearchGraphBuilder from the main app
# (Include the full class here from the previous artifact)
# For brevity, I'll include a placeholder - in production, include the full class

class EnhancedResearchGraphBuilder:
    """Enhanced graph builder with all the features from previous implementation"""
    def __init__(self):
        self.graph = nx.Graph()
        self.nodes = []
        self.edges = []
        self.node_types = {}
        self.communities = {}
        
    def process_excel_file(self, filepath: str) -> Dict[str, Any]:
        """Process Excel file and build graph - simplified for example"""
        try:
            df = pd.read_excel(filepath)
            # Process data (simplified for production template)
            for idx, row in df.iterrows():
                paper_id = f"paper_{idx}"
                self.nodes.append({
                    'id': paper_id,
                    'label': f"Paper {idx + 1}",
                    'type': 'paper',
                    'group': 'papers',
                    'size': 10
                })
            
            return self._format_graph_data()
        except Exception as e:
            logger.error(f"Error processing file: {str(e)}")
            raise
    
    def _format_graph_data(self) -> Dict[str, Any]:
        return {
            'nodes': self.nodes,
            'links': self.edges,
            'metadata': {
                'total_nodes': len(self.nodes),
                'total_edges': len(self.edges),
                'generated_at': datetime.now().isoformat()
            }
        }

# Initialize graph builder
graph_builder = EnhancedResearchGraphBuilder()

# Health check endpoint for Render
@app.route('/api/health')
def health_check():
    """Health check endpoint for monitoring"""
    return jsonify({
        'status': 'healthy',
        'timestamp': datetime.now().isoformat(),
        'version': '1.0.0'
    }), 200

@app.route('/')
def index():
    """Serve the main application page"""
    try:
        return render_template('index.html')
    except Exception as e:
        logger.error(f"Error serving index: {str(e)}")
        return jsonify({'error': 'Template not found'}), 500

@app.route('/api/upload', methods=['POST'])
def upload_file():
    """Handle file upload with production error handling"""
    try:
        if 'file' not in request.files:
            return jsonify({'error': 'No file provided'}), 400
        
        file = request.files['file']
        if file.filename == '':
            return jsonify({'error': 'No file selected'}), 400
        
        if file and allowed_file(file.filename):
            filename = secure_filename(file.filename)
            filepath = os.path.join(app.config['UPLOAD_FOLDER'], filename)
            file.save(filepath)
            
            # Process the file
            graph_data = graph_builder.process_excel_file(filepath)
            
            # Save processed data
            output_path = os.path.join('static', 'data', 'graph_data.json')
            with open(output_path, 'w') as f:
                json.dump(graph_data, f, indent=2)
            
            # Clean up uploaded file to save space
            if os.path.exists(filepath):
                os.remove(filepath)
            
            return jsonify({
                'success': True,
                'message': 'File processed successfully',
                'graph_data': graph_data
            })
        
        return jsonify({'error': 'Invalid file type'}), 400
        
    except Exception as e:
        logger.error(f"Upload error: {str(e)}")
        return jsonify({'error': 'Processing failed'}), 500

@app.route('/api/graph')
def get_graph():
    """Get current graph data with caching headers"""
    try:
        graph_path = os.path.join('static', 'data', 'graph_data.json')
        if os.path.exists(graph_path):
            with open(graph_path, 'r') as f:
                graph_data = json.load(f)
            
            response = jsonify(graph_data)
            response.headers['Cache-Control'] = 'public, max-age=300'  # 5 min cache
            return response
        else:
            return jsonify(get_sample_graph_data()), 200
            
    except Exception as e:
        logger.error(f"Error getting graph: {str(e)}")
        return jsonify({'error': 'Failed to load graph'}), 500

@app.route('/api/search', methods=['POST'])
def search_graph():
    """Search nodes in the graph"""
    try:
        query = request.json.get('query', '').lower()
        graph_path = os.path.join('static', 'data', 'graph_data.json')
        
        if os.path.exists(graph_path):
            with open(graph_path, 'r') as f:
                graph_data = json.load(f)
            
            matching_nodes = []
            for node in graph_data.get('nodes', []):
                if query in str(node).lower():
                    matching_nodes.append(node['id'])
            
            return jsonify({
                'results': matching_nodes[:100],  # Limit results
                'count': len(matching_nodes)
            })
        
        return jsonify({'results': [], 'count': 0})
        
    except Exception as e:
        logger.error(f"Search error: {str(e)}")
        return jsonify({'error': 'Search failed'}), 500

@app.route('/api/data/clear', methods=['POST'])
def clear_data():
    """Clear all data with backup"""
    try:
        graph_path = os.path.join('static', 'data', 'graph_data.json')
        if os.path.exists(graph_path):
            # Create backup
            timestamp = datetime.now().strftime('%Y%m%d_%H%M%S')
            backup_file = f'backups/backup_{timestamp}.json'
            os.makedirs('backups', exist_ok=True)
            
            with open(graph_path, 'r') as f:
                data = json.load(f)
            with open(backup_file, 'w') as f:
                json.dump(data, f)
            
            os.remove(graph_path)
            
            return jsonify({
                'success': True,
                'message': 'Data cleared',
                'backup': backup_file
            })
        
        return jsonify({'success': True, 'message': 'No data to clear'})
        
    except Exception as e:
        logger.error(f"Clear error: {str(e)}")
        return jsonify({'error': 'Failed to clear data'}), 500

@app.route('/api/data/stats', methods=['GET'])
def get_stats():
    """Get graph statistics"""
    try:
        graph_path = os.path.join('static', 'data', 'graph_data.json')
        if os.path.exists(graph_path):
            with open(graph_path, 'r') as f:
                data = json.load(f)
            
            return jsonify({
                'total_nodes': len(data.get('nodes', [])),
                'total_edges': len(data.get('links', [])),
                'metadata': data.get('metadata', {})
            })
        
        return jsonify({'total_nodes': 0, 'total_edges': 0})
        
    except Exception as e:
        logger.error(f"Stats error: {str(e)}")
        return jsonify({'error': 'Failed to get stats'}), 500

def allowed_file(filename):
    """Check if file extension is allowed"""
    return '.' in filename and \
           filename.rsplit('.', 1)[1].lower() in app.config['ALLOWED_EXTENSIONS']

def get_sample_graph_data():
    """Return sample graph data"""
    return {
        'nodes': [
            {'id': 'welcome', 'label': 'Upload your Excel file to begin', 
             'type': 'info', 'size': 20}
        ],
        'links': [],
        'metadata': {'total_nodes': 1, 'total_edges': 0}
    }

@app.errorhandler(404)
def not_found(error):
    """Handle 404 errors"""
    if request.path.startswith('/api/'):
        return jsonify({'error': 'Endpoint not found'}), 404
    return render_template('index.html'), 200

@app.errorhandler(500)
def internal_error(error):
    """Handle 500 errors"""
    logger.error(f"Internal error: {str(error)}")
    return jsonify({'error': 'Internal server error'}), 500

# Add security headers
@app.after_request
def add_security_headers(response):
    """Add security headers to all responses"""
    response.headers['X-Content-Type-Options'] = 'nosniff'
    response.headers['X-Frame-Options'] = 'SAMEORIGIN'
    response.headers['X-XSS-Protection'] = '1; mode=block'
    response.headers['Strict-Transport-Security'] = 'max-age=31536000; includeSubDomains'
    return response

if __name__ == '__main__':
    port = int(os.getenv('PORT', 5000))
    debug = os.getenv('FLASK_ENV') == 'development'
    app.run(host='0.0.0.0', port=port, debug=debug)