requirements_prod.txt
# Production dependencies for Research Knowledge Graph

# Core Flask Framework
Flask==3.0.0
Flask-CORS==4.0.0

# WSGI Server for production
gunicorn==21.2.0

# Data Processing
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2  # Excel file support
xlrd==2.0.1      # Legacy Excel format support

# Graph Processing
networkx==3.2.1

# Utilities
Werkzeug==3.0.1
python-dotenv==1.0.0
python-dateutil==2.8.2

# Security
itsdangerous==2.1.2  # For secure sessions
cryptography==41.0.7  # For secure data handling

# Performance & Monitoring (optional but recommended)
flask-compress==1.14  # Gzip compression
flask-limiter==3.5.0  # Rate limiting
redis==5.0.1         # For rate limiting backend

# Logging
python-json-logger==2.0.7  # Structured logging

# File handling
python-magic==0.4.27  # Better file type detection

# Optional: Enhanced NLP for concept extraction
# Uncomment if you want advanced text processing
# spacy==3.7.2
# scikit-learn==1.3.2

# Optional: Database support for future
# SQLAlchemy==2.0.23
# psycopg2-binary==2.9.9